requirements.txt
# Production dependencies for TraderAgent Docker container
# Core dependencies
openai>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0

# Additional dependencies for robust deployment
python-dotenv>=1.0.0  # For environment variable management
urllib3>=2.0.0        # HTTP client (often required by requests)
certifi>=2023.0.0     # SSL certificates
charset-normalizer>=3.0.0  # Character encoding detection

# Optional: For enhanced error handling and logging
packaging>=23.0       # Version parsing utilities